requirements.txt
Flask==0.10.1
Flask-PyMongo==0.3.1
unicode-slugify==0.1.3
requests==1.2.3
Flask-Cache==0.13.1
futures==3.0.5
funcsigs==0.4
apscheduler==3.3.1
Flask-Cors==3.0.2